Major Bug Fixes

+++ src/bitset/shifts.rs
@@ -2,11 +2,11 @@ use crate::{BitSet, Cols};
 
 impl<const T: usize> BitSet<T> {
     pub fn up(&self, shift: u32, cols: Cols) -> BitSet<T> {
-        *self >> shift * (cols as u32)
+        *self >> shift * (cols)
     }
 
     pub fn down(&self, shift: u32, cols: Cols) -> BitSet<T> {
-        *self << shift * (cols as u32)
+        *self << shift * (cols)
     }
 
     pub fn right(&self, shift: u32) -> BitSet<T> {
@@ -18,11 +18,11 @@ impl<const T: usize> BitSet<T> {
     }
 
     pub fn up_mut(&mut self, shift: u32, cols: Cols) {
-        *self >>= shift * (cols as u32);
+        *self >>= shift * (cols);
     }
 
     pub fn down_mut(&mut self, shift: u32, cols: Cols) {
-        *self <<= shift * (cols as u32);
+        *self <<= shift * (cols);
     }
 
     pub fn right_mut(&mut self, shift: u32) {
